Added support for `lower_bound`
The trace function now returns position of key (mis)match (with match
indication).
That may be provided back to the user; upon mismatch, the item
may be inserted to the lower bound position in O(1) time (allowing for
efficient find-or-insert operation).

Note that the `lower_bound` function is not tested, yet.